requirements.txt
astroid==2.15.6
behave==1.2.6
black==23.7.0
click==8.1.7
dill==0.3.7
importlib-metadata==6.8.0
isort==5.12.0
joblib==1.3.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
parse==1.20.1
parse-type==0.6.2
pathspec==0.11.2
platformdirs==3.10.0
psycopg==3.1.10
psycopg-pool==3.1.7
psycopg2==2.9.7
PyHamcrest==2.1.0
pylint==2.17.5
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typing_extensions==4.7.1
wrapt==1.15.0
yoyo-migrations==8.2.0
zipp==3.17.0
